pyproject.toml
[tool.poetry]
name = "expense-api-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
asyncpg = "^0.29.0"
uvicorn = "^0.31.1"
alembic = "^1.13.3"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
pydantic = "^2.9.2"
greenlet = "^3.1.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
result = "^0.17.0"
pytest-asyncio = "^0.24.0"
pyre-check = "^0.9.22"
sqlalchemy-utils = "^0.41.2"
factory-boy = "^3.3.1"
cryptography = "^43.0.1"
returns = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
